Moving BP files to new directory
