+++ src/include/simeng/arch/aarch64/Instruction.hh
@@ -3,7 +3,7 @@
 #include <array>
 #include <unordered_map>
 
-#include "simeng/BranchPredictor.hh"
+#include "simeng/branchPredictors/BranchPredictor.hh"
 #include "simeng/Instruction.hh"
 #include "simeng/arch/aarch64/InstructionGroups.hh"
 #include "simeng/arch/aarch64/operandContainer.hh"
