+++ src/include/simeng/arch/riscv/Instruction.hh
@@ -5,7 +5,7 @@
 #include <functional>
 #include <unordered_map>
 
-#include "simeng/BranchPredictor.hh"
+#include "simeng/branchPredictors/BranchPredictor.hh"
 #include "simeng/Instruction.hh"
 #include "simeng/arch/riscv/InstructionGroups.hh"
