+++ src/lib/CMakeLists.txt
@@ -14,6 +14,8 @@ set(SIMENG_SOURCES
     arch/riscv/Instruction_decode.cc
     arch/riscv/Instruction_execute.cc
     arch/riscv/InstructionMetadata.cc
+    branchPredictors/GenericPredictor.cc
+    branchPredictors/PerceptronPredictor.cc
     config/ModelConfig.cc
     kernel/Linux.cc
     kernel/LinuxProcess.cc
@@ -40,8 +42,6 @@ set(SIMENG_SOURCES
     CMakeLists.txt
     CoreInstance.cc
     Elf.cc
-    GenericPredictor.cc
-    PerceptronPredictor.cc
     RegisterFileSet.cc
     RegisterValue.cc
     SpecialFileDirGen.cc
